+++ metadata.rb
@@ -6,7 +6,7 @@ maintainer_email 'peter@realityforge.org'
 license 'Apache 2.0'
 description 'Installs/Configures kibana, the logstash UI'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '0.1.11'
+version '0.1.12'
 
 supports 'ubuntu'
 supports 'debian'