+++ recipes/_service.rb
@@ -1,11 +1,5 @@
 # Encoding: utf-8
 
-service 'kibana' do
-  provider Chef::Provider::Service::Upstart
-  supports start: true, restart: true, stop: true, status: true
-  action :nothing
-end
-
 template '/etc/init/kibana.conf' do
   cookbook node['kibana']['service']['cookbook']
   source node['kibana']['service']['source']
@@ -15,4 +9,16 @@ template '/etc/init/kibana.conf' do
     recent_upstart: (node['platform_family'] != 'rhel')
   )
   notifies :restart, 'service[kibana]', :delayed
+  not_if KibanaUtils.upstart_supports?(node)
 end
+
+service 'kibana' do
+  if KibanaUtils.upstart_supports?(node)
+    provider Chef::Provider::Service::Upstart
+  else
+    provider Chef::Provider::Service::Systemd
+  end
+  supports start: true, restart: true, stop: true, status: true
+  action :nothing
+end
+